refactor(workspace): extract useWorkspaceSetup hook (1646→1256 lines)
Moves practice management, onboarding conversation init, setup save handlers,
logo upload, Stripe status, and billing data into a dedicated hook.
